refactor out all backend interaction to client.py
